Refactoring after new database

+++ cdr-processor/src/main/java/net/gcicom/cdr/processor/repository/CDRAuditRepository.java
@@ -1,9 +1,12 @@
 package net.gcicom.cdr.processor.repository;
 
+import java.math.BigInteger;
+
 import org.springframework.data.repository.CrudRepository;
 
-import net.gcicom.cdr.processor.entity.audit.AuditEvent;
+import net.gcicom.domain.imported.events.AuditEvent;
 
-public interface CDRAuditRepository extends CrudRepository<AuditEvent, Long> {
+
+public interface CDRAuditRepository extends CrudRepository<AuditEvent, BigInteger> {
 
 }
